refactor components; lifting state up
