+++ src/components/Card.js
@@ -1,32 +1,12 @@
-import { Card as ECard, Checkbox, Pane, RadioGroup } from 'evergreen-ui';
+import { Card as ECard, Pane } from 'evergreen-ui';
 import QRCode from 'qrcode.react';
-import { useEffect, useRef, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import './style.css';
 
-export const Card = ({ direction = 'ltr' }) => {
+export const Card = (props) => {
   const { t } = useTranslation();
-  const firstLoad = useRef(true);
   const [qrvalue, setQrvalue] = useState('');
-  const [settings, setSettings] = useState({
-    // Network SSID name
-    ssid: '',
-    // Network password
-    password: '',
-    // Show advanced options
-    showAdvanced: false,
-    // Advanced option: Network encryption mode
-    encryptionMode: 'WPA',
-    // Advanced option: Hide password on the printed card
-    hidePassword: false,
-    // Advanced option: Portrait orientation
-    portrait: false,
-  });
-  const [encryptionModes] = useState([
-    { label: 'None', value: 'nopass' },
-    { label: 'WPA/WPA2/WPA3', value: 'WPA' },
-    { label: 'WEP', value: 'WEP' },
-  ]);
 
   const escape = (v) => {
     const needsEscape = ['"', ';', ',', ':', '\\'];
@@ -42,47 +22,16 @@ export const Card = ({ direction = 'ltr' }) => {
     return escaped;
   };
 
-  const onPrint = () => {
-    if (settings.ssid.length > 0) {
-      if (settings.password.length < 8 && settings.encryptionMode === 'WPA') {
-        alert(t('wifi.alert.password.length.8'));
-      } else if (
-        settings.password.length < 5 &&
-        settings.encryptionMode === 'WEP'
-      ) {
-        alert(t('wifi.alert.password.length.5'));
-      } else {
-        document.title = 'WiFi Card - ' + settings.ssid;
-        window.print();
-      }
-    } else {
-      alert(t('wifi.alert.name'));
-    }
-  };
-
-  const disableHidePassword = () => {
-    const isWEPWithPasswordLengthShorterThat5Characters = () => {
-      return settings.encryptionMode === 'WEP' && settings.password.length < 5
-        ? true
-        : false;
-    };
-
-    return settings.encryptionMode === 'WPA' && settings.password.length < 8
-      ? true
-      : isWEPWithPasswordLengthShorterThat5Characters();
-  };
-
   useEffect(() => {
-    if (firstLoad.current && window.innerWidth < 500) {
-      firstLoad.current = false;
-      setSettings({ ...settings, portrait: true });
-    }
-
-    const ssid = escape(settings.ssid);
+    const ssid = escape(props.settings.ssid);
     const password =
-      settings.encryptionMode === 'nopass' ? '' : escape(settings.password);
-    setQrvalue(`WIFI:T:${settings.encryptionMode};S:${ssid};P:${password};;`);
-  }, [settings]);
+      props.settings.encryptionMode === 'nopass'
+        ? ''
+        : escape(props.settings.password);
+    setQrvalue(
+      `WIFI:T:${props.settings.encryptionMode};S:${ssid};P:${password};;`
+    );
+  }, [props.settings]);
 
   const setEncryptionMode = (e) =>
     setNetwork({
@@ -98,21 +47,21 @@ export const Card = ({ direction = 'ltr' }) => {
       <ECard
         id="print-area"
         elevation={3}
-        style={{ maxWidth: settings.portrait ? '350px' : '100%' }}
+        style={{ maxWidth: props.settings.portrait ? '350px' : '100%' }}
       >
-        <h1 style={{ textAlign: settings.portrait ? 'center' : 'unset' }}>
+        <h1 style={{ textAlign: props.settings.portrait ? 'center' : 'unset' }}>
           {t('wifi.login')}
         </h1>
 
         <div
           className="details"
-          style={{ flexDirection: settings.portrait ? 'column' : 'row' }}
+          style={{ flexDirection: props.settings.portrait ? 'column' : 'row' }}
         >
           <QRCode
             className="qrcode"
             style={
-              !settings.portrait
-                ? direction === 'ltr'
+              !props.settings.portrait
+                ? props.direction === 'ltr'
                   ? { paddingRight: '1em' }
                   : { paddingLeft: '1em' }
                 : {}
@@ -132,15 +81,13 @@ export const Card = ({ direction = 'ltr' }) => {
               autoCorrect="off"
               autoCapitalize="none"
               spellCheck="false"
-              value={settings.ssid}
-              onChange={(e) =>
-                setSettings({ ...settings, ssid: e.target.value })
-              }
+              value={props.settings.ssid}
+              onChange={(e) => props.onSSIDChange(e.target.value)}
             />
             <label
               className={`
-                ${settings.hidePassword && 'no-print hidden'}
-                ${settings.encryptionMode === 'nopass' && 'hidden'}
+                ${props.settings.hidePassword && 'no-print hidden'}
+                ${props.settings.encryptionMode === 'nopass' && 'hidden'}
               `}
             >
               {t('wifi.password')}
@@ -149,12 +96,12 @@ export const Card = ({ direction = 'ltr' }) => {
               id="password"
               type="text"
               className={`
-                ${settings.hidePassword && 'no-print hidden'}
-                ${settings.encryptionMode === 'nopass' && 'hidden'}
+                ${props.settings.hidePassword && 'no-print hidden'}
+                ${props.settings.encryptionMode === 'nopass' && 'hidden'}
               `}
               style={{
                 height:
-                  settings.portrait && settings.password.length > 40
+                  props.settings.portrait && props.settings.password.length > 40
                     ? '5em'
                     : 'auto',
               }}
@@ -164,10 +111,8 @@ export const Card = ({ direction = 'ltr' }) => {
               autoCorrect="off"
               autoCapitalize="none"
               spellCheck="false"
-              value={settings.password}
-              onChange={(e) => {
-                setSettings({ ...settings, password: e.target.value });
-              }}
+              value={props.settings.password}
+              onChange={(e) => props.onPasswordChange(e.target.value)}
             />
           </div>
         </div>
@@ -179,45 +124,6 @@ export const Card = ({ direction = 'ltr' }) => {
           {t('wifi.tip')}
         </p>
       </ECard>
-
-      <div
-        id="settings"
-        style={{ maxWidth: settings.portrait ? '350px' : '100%' }}
-      >
-        <RadioGroup
-          label={`${t('wifi.password.encryption')}:${
-            direction === 'rtl' ? ' ' : ''
-          }`}
-          size={16}
-          value={settings.encryptionMode}
-          options={encryptionModes}
-          onChange={(e) =>
-            setSettings({ ...settings, encryptionMode: e.target.value })
-          }
-        />
-
-        <Checkbox
-          label={t('wifi.password.hide')}
-          checked={settings.hidePassword}
-          disabled={disableHidePassword()}
-          onChange={() =>
-            setSettings({ ...settings, hidePassword: !settings.hidePassword })
-          }
-        />
-        <Checkbox
-          label={t('button.rotate')}
-          checked={settings.portrait}
-          onChange={() =>
-            setSettings({ ...settings, portrait: !settings.portrait })
-          }
-        />
-      </div>
-
-      <div className="buttons">
-        <button id="print" onClick={onPrint}>
-          {t('button.print')}
-        </button>
-      </div>
     </Pane>
   );
 };
